Add files via upload
Write a simple English <> Native language (Korean or French) translation program. Assign at least 10 English words to the array.
String eng[] = {“apple”, “is”, “red”, ..., XX}
String kor[] = {“사과”, “는”, “빨간색”, ... , XXX}
Based on this information, write a program that translates when the user enters a sentence (assuming a sentence is entered within the assigned English word). If the word is not assigned in the program, print "I don't know".